+++ src/main/php/lang/reflection/Constant.class.php
@@ -6,6 +6,9 @@ class Constant extends Member {
 
   protected function getAnnotations() { return Reflection::parse()->ofConstant($this->reflect); }
 
+  /** Returns a compound name consisting of `[CLASS]::$[NAME]`  */
+  public function compoundName(): string { return strtr($this->reflect->class, '\\', '.').'::'.$this->reflect->name; }
+
   /** @return int */
   public function modifiers() { return $this->reflect->getModifiers(); }