+++ src/main/php/lang/reflection/Method.class.php
@@ -39,7 +39,7 @@ class Method extends Routine {
     } catch (\ReflectionException $e) {
       throw new CannotInvoke(strtr($this->reflect->class, '\\', '.').'::'.$this->reflect->name, $e);
     } catch (\Throwable $e) {
-      throw new InvocationFailed(strtr($this->reflect->class, '\\', '.').'::'.$this->reflect->name, $e);
+      throw new InvocationFailed($this, $e);
     }
   }