+++ ChangeLog.md
@@ -3,6 +3,12 @@ XP Reflection ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added `compoundName()` accessor to constants, properties and methods.
+  (@thekid)
+* Passed method or constructor to `InvocationFailed` exception, and make
+  it accessible via its `target()` method.
+  (@thekid)
+
 ## 0.4.0 / 2020-12-14
 
 * Removed lookup for lowercase XP annotations, the AST library does not